Moved job related method into JobController from rest api
